fix(cli): fix windows is_process_running type
Fix Windows CI compilation by casting `STILL_ACTIVE` to u32 when comparing with the `GetExitCodeProcess` exit code.

+++ crates/bifrost-cli/src/process.rs
@@ -119,7 +119,7 @@ pub fn is_process_running(pid: u32) -> bool {
         CloseHandle(handle);
     }
 
-    ok != 0 && exit_code == STILL_ACTIVE
+    ok != 0 && exit_code == STILL_ACTIVE as u32
 }
 
 #[cfg(all(not(unix), not(windows)))]
